Split last “Multiple types” example to replace the entier guide example via the select
